content/canvas/test/test_2d.imageData.put.type.onepixel.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.imageData.put.type.onepixel</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.putImageData({ width: 1, height: 1, data: [0, 0, 0, 0] }, 0, 0);
ok(true, "true"); // no exception

if (!_deferred) SimpleTest.finish();
});
</script>
